+++ example/lib/main.dart
@@ -43,7 +43,7 @@ class MyApp extends StatelessWidget {
                     errorStyle: TextStyle(color: Colors.red),
                     hintStyle: TextStyle(color: Colors.grey),
                   ),
-                  dropdownBoxDecoration:  DropDownBoxDecoration(
+                  dropdownBoxDecoration: DropDownBoxDecoration(
                     color: Colors.white,
                     shape: RoundedRectangleBorder(
                       borderRadius: BorderRadius.circular(10),
@@ -63,11 +63,10 @@ class MyApp extends StatelessWidget {
                   },
                   onChangedSearch: (searchTerm) async {
                     // Replace this with your own search logic
-                    return ['Item1', 'Item2', 'Item3'];
+                    return ['Item1 sssssssssssss', 'Item2', 'Item3'];
                   },
-                  dropdownItems:const  ['Item1', 'Item2', 'Item3'],
+                  dropdownItems:const  ['Item1 sssssssssssss', 'Item2', 'Item3'],
                   // hint: 'Select an item',
-                  initValue: 'Item1',
                   // buttonWidth: 200,
                   itemHeight: 50,
                   dropdownHeight: 200,
@@ -81,6 +80,36 @@ class MyApp extends StatelessWidget {
                     errorStyle: TextStyle(color: Colors.red),
                     hintStyle: TextStyle(color: Colors.grey),
                   ),
+                  itemBuilder: (item, isSelected) {
+                    return Container(
+                      alignment: Alignment.centerLeft,
+                      padding: const EdgeInsets.symmetric(vertical: 10, horizontal: 5),
+                      decoration: BoxDecoration(
+                        color: isSelected ? Colors.red.withOpacity(0.1) : Colors.transparent,
+                        borderRadius: BorderRadius.circular(10),
+                      ),
+                      child: Row(
+                        children: [
+                          if (isSelected) ...[
+                            const Icon(
+                              Icons.check,
+                              size: 16,
+                              color: Color(0xff111111),
+                            ),
+                            const SizedBox(width: 7.5),
+                          ],
+                          Expanded(
+                            child: Text(
+                              item,
+                              maxLines: 3,
+                              style: Theme.of(context).textTheme.labelMedium!.copyWith(
+                                  color: const Color(0xff111111), fontWeight: FontWeight.w400),
+                            ),
+                          ),
+                        ],
+                      ),
+                    );
+                  },
                   buttonDecoration: const InputDecoration(
                     hintText: 'Select an item',
                     labelText: 'Search',