+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to the "magic_dropdown_search" project will be documented in
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.8] - 2024-06-13
+
+### Added
+- Added the `itemBuilder` property to customize the dropdown items.
+
+### Fixed
+- Fixed the issue with the hint style.
+- Fixed the issue with the dropdown style not updating.
+
+
 ## [0.0.7] - 2024-06-12
 
 ### Fixed